+++ packages/lexicon/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v1.2.6
+
+- chore: regenerated from synced lexicons
+
 ## v1.2.5
 
 - chore: regenerated from synced lexicons